requirements.txt
numpy==1.19.2
Pillow==8.4.0
torch==1.10.2
torchvision==0.11.3
scikit-learn==0.24.2
pandas==1.1.5
seaborn==0.11.2